PDM2D/requirements.txt
PySide6>=6.5.0
python-dotenv>=1.0.0